Fundamental memory model refactor
Rather than storing all the game RAM into a massive C-Style array, we can access a C++ data structure which will allow us to implement some greater control logic. E.g., overriding accesses to certain regions

Let's say something asks for the memory in address 0x20. We can now, with this new structure, optionally check some external flag which can serve as a kind of 'virtual multiplexer' input, and say if that flag is set, we pull from data source A, if not, pull from data source B. All the while without serious modification to the C code that wants this RAM data.

The hope is that this will be of some help in adding local multiplayer